OME-Zarr: Extract connection logic
It's about to get even more complex
